pyproject.toml
[tool.poetry]
name = "mdfb"
version = "1.6.1"
description = "A CLI for downloading posts in bulk from Bluesky from a specified account"
authors = ["Ibrahim Haji Abdi <ibrahim.hajiabdi09@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
mdfb = "mdfb.mdfb:main"

[tool.poetry.dependencies]
python = ">3.9.1,<3.14"
atproto = "^0.0.55"
tdqm = "^0.0.1"
argparse = "^1.4.0"
pathvalidate = "^3.2.1"
tqdm = "^4.67.1"
tenacity = "^9.0.0"
platformdirs = "^4.3.7"
h11 = "^0.16.0"
httpcore = "^1.0.9"
psutil = "^7.0.0"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
requests-mock = "^1.12.1"
pytest-mock = "^3.14.0"
pytest = "^8.3.4"

[project]
name = "mdfb"
version = "1.6.1"
authors = [
  { name="Ibrahim Haji Abdi", email="ibrahim.hajiabdi09@gmail.com" },
]
readme = "README.md"
requires-python = ">3.9.1,<3.14"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Environment :: Console"
]

[tool.poetry.urls]
Repository = "https://github.com/IbrahimHajiAbdi/mass-downloader-for-bluesky.git"
Issues = "https://github.com/IbrahimHajiAbdi/mass-downloader-for-bluesky/issues"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
